license_portal/requirements.txt
celery==5.3.1
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
djangorestframework==3.14.0
psycopg2-binary==2.9.7
